requirements.txt
verbecc
wordfreq
deep-translator
streamlit>=1.32
rapidfuzz>=3.0
st-gsheets-connection
pandas
streamlit-keyup
requests
fastapi
uvicorn
sqlalchemy
authlib
itsdangerous
email-validator
python-dotenv
